requirements.txt
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
requests
openai


python-dotenv
